adapt readme and store script in new folder
